inet_addr(), rand(), unit tests updated
Switched code from inet_aton to inet_addr
Switched code from random() to rand()
Unit tests now compiles and runs well
Updated the queue test
Fixed a bug in the queue module

+++ tests/memcpy.c
@@ -5,8 +5,8 @@ int main(int argc, char *argv[])
     packet_t packet;
     char ppacket[PACKET_SIZE];
 
-    printf("&ppacket = %i\n", &ppacket);
-    printf("&ppacket + sizeof(packet.id) = %i\n",
-           (char *) &ppacket + sizeof(packet.id));
+    printf("&ppacket = %u\n", (int) &ppacket);
+    printf("&ppacket + sizeof(packet.id) = %u\n",
+           (int) ((char *) &ppacket + sizeof(packet.id)));
     return 0;
 }
